create separate folders for agent/appsync resolver tests
