+++ packages/event-handler/tests/unit/appsync-events/AppSyncEventsResolver.test.ts
@@ -3,11 +3,11 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import {
   AppSyncEventsResolver,
   UnauthorizedException,
-} from '../../src/appsync-events/index.js';
+} from '../../../src/appsync-events/index.js';
 import {
   onPublishEventFactory,
   onSubscribeEventFactory,
-} from '../helpers/factories.js';
+} from '../../helpers/factories.js';
 
 describe('Class: AppSyncEventsResolver', () => {
   beforeEach(() => {
